requirements.txt
asgiref==3.9.2
Django==4.2.24
django-admin==2.0.2
django-cors-headers==4.9.0
django-excel-response2==3.0.6
django-rest-framework==0.1.0
django-six==1.0.5
djangorestframework==3.16.1
excel-base==1.0.4
isoweek==1.3.3
python-dateutil==2.9.0.post0
screen==1.0.1
six==1.17.0
sqlparse==0.5.3
TimeConvert==3.0.13
typing_extensions==4.15.0
tzlocal==5.3.1
xlwt==1.3.0
